% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmw_benchmarks_2025}
\alias{bmw_benchmarks_2025}
\title{BMW brand benchmarks for 2025}
\format{
A data frame.
}
\source{
Internal benchmark workbook.
}
\usage{
bmw_benchmarks_2025
}
\description{
Data frames containing 2025 benchmarks for all metrics, attributes, and traits for BMW
}
\details{
Data frames can be found for the Overall audience, Women, AAPI, and Gen Z/Millennials.
}
\keyword{datasets}
