% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creative_benchmarks_2025}
\alias{creative_benchmarks_2025}
\title{Creative benchmarks for 2025}
\format{
A data frame.
}
\source{
Internal benchmark workbook.
}
\usage{
creative_benchmarks_2025
}
\description{
Data frames containing 2025 creative benchmark values used in the brand
tracker. Benchark values are split out by "Social" and "Other".
}
\details{
Data frames can be found for the Overall audience, Women, AAPI, and Gen Z/Millennials.
}
\keyword{datasets}
