% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmw_benchmarks.R
\name{get_bmw_benchmarks}
\alias{get_bmw_benchmarks}
\title{Get BMW benchmark data}
\usage{
get_bmw_benchmarks(type = c("overall", "quarter"), year = "2025")
}
\arguments{
\item{type}{Type of benchmark to return. \code{"overall"} for annual benchmarks,
\code{"quarter"} for quarterly benchmarks.}

\item{year}{Benchmark year (character or numeric).}
}
\value{
A data frame of BMW benchmark data.
}
\description{
Returns BMW benchmark data, either annual or quarterly, for a given year.
}
