% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creative_benchmarks.R
\name{get_creative_benchmarks}
\alias{get_creative_benchmarks}
\title{Get creative benchmark data}
\usage{
get_creative_benchmarks()
}
\value{
A data frame of creative benchmark values.
}
\description{
Returns the packaged creative benchmark dataset.
}
