% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSPtracker-package.R
\docType{package}
\name{GSPtracker-package}
\alias{GSPtracker}
\alias{GSPtracker-package}
\title{GSPtracker: Process Lucid/Cint tracker data}
\description{
A package to easily process Cint tracker data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/taylorgrant/GSPtracker}
  \item Report bugs at \url{https://github.com/taylorgrant/GSPtracker/issues}
}

}
\author{
\strong{Maintainer}: Taylor Grant \email{taylor.grant@omc.com}

}
\keyword{internal}
