% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_summary.R
\name{brand_summary}
\alias{brand_summary}
\title{Summarize tracker data for a specific brand}
\usage{
brand_summary(
  data,
  groups = NULL,
  qq,
  include_month = FALSE,
  include_quarter = FALSE,
  moving_average = FALSE
)
}
\arguments{
\item{data}{Tracker dataset to use}

\item{groups}{Group to cut by, or NULL}

\item{qq}{Tracker question of interest}

\item{include_month}{Logical - break data out by month of tracker}

\item{include_quarter}{Logical - roll data up to quarters}

\item{moving_average}{Logical - include 3 month rolling average}
}
\value{
Dataframe
}
\description{
Summarizes the control/exposed across brand metrics, brand traits, and brand attributes. If using
the unweighted data set, it returns total population percentages.
}
\examples{
\dontrun{
# individual question
brand_summary(cint$campaign, groups = NULL, qq = "unaided_awareness_client")
# multiple questions (using map::purrr)
c("unaided_awareness_client", "awr_a_1", "awr_aad_5") |>
purrr::map_dfr(~ brand_summary(data = cint$unweighted, groups = NULL, qq = .x))
}
}
