% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cint.R
\name{read_cint}
\alias{read_cint}
\title{Read in tracker data from Cint}
\usage{
read_cint(file_loc)
}
\arguments{
\item{file_loc}{Give the location of the tracker file}
}
\value{
Named list. Raw datafile (df), all others are in srvyr survey format
}
\description{
This function reads in the raw data and weights the data using the weights provided by Cint in the file.
}
\examples{
\dontrun{
file_loc <- "/folder/director/file.csv"
cint <- read_cint(file_loc)
}
}
